01-Q1: Add the implementation of quiz 1
[Description]
Consider that you have a rectangular piece of paper of arbitrary
dimensions N by M (where N and M are positive integers). You also
have a pair of scissors which can cut perfectly straight with no
loss of paper. You wish to reduce the original piece of paper into
a series of perfect squares of paper, making the largest possible
squares, and using all of the paper provided. Write a function in
C that takes the inputs N, and M, and returns the series of squares
that can be made out of that piece of paper. No fractional squares,
i.e. no square should be less than 1 in length and width.

[Test]
N=6, M=5
N=1, M=1
N=9, M=9