01-Q2: Add the implementation of quiz 2
[Description]
a) Find function f(Ai) using these sets of points in the spreadsheet.
b) What would be the output for the following inputs?
	* f(30001) = ?
	* f(55555) = ?
	* f(77788) = ?
c) What is the upper limit or maximum range of this function before
   there will be collisions or overflow?
d) Please share any code that you write in the process of solving
   the above problems.

a) f(Ai) is a function that convert decimal numbers into 62-based
   numbers where the digits and alphabetic characters are rearranged,
   which means the next number of 5 may not be 6.
   The mapping table is as below,
	0=C,    1=7,    2=x,    3=i,    4=c,    5=P,    6=M,    7=G,
	8=v,    9=z,    10=A,   11=Z,   12=y,   13=T,   14=N,   15=o,
	16=d,   17=m,   18=w,   19=n,   20=V,   21=5,   22=D,   23=3,
	24=B,   25=6,   26=H,   27=0,   28=O,   29=u,   30=p,   31=8,
	32=E,   33=2,   34=1,   35=W,   36=9,   37=s,   38=q,   39=L,
	40=Q,   41=X,   42=4,   43=Y,   44=j,   45=S,   46=e,   47=U,
	48=h,   49=I,   50=R,   51=J,   52=g,   53=a,   54=f,   55=F,
	56=t,   57=r,   58=K,   59=l,   60=b,   61=k,
b) f(30001) = GIF
   f(55555) = NOi
   f(77788) = VNQ
c) There is no upper limit or maximum range in this function.
d) Please reference the source code

[Test]
Test set 1:
All the numbers in the original spreadsheet.

Test set 2:
30001
55555
77788